Move tested chat test in test_nested
